move MultiCurrencyAdapter to traits & decrease crowdloans' MigrateKeysLimit
