+++ docker-compose.override.yml
@@ -14,10 +14,6 @@ services:
     image: parallelfinance/oracle-client:latest
     command: start -m 12000 -w ws://parachain-${PARA_ID}-0:9944
     restart: always
-  nominate-client:
-    image: parallelfinance/nominate-client:latest
-    command: --relay-ws ws://relaychain-alice:9944 --para-ws ws://parachain-${PARA_ID}-0:9944 --tick 120000
-    restart: always
   stake-client-claim:
     image: parallelfinance/stake-client:latest
     command: claim --para-ws=ws://parachain-${PARA_ID}-0:9944 --batch-size 50
