+++ pallets/xcm-helper/src/mock.rs
@@ -12,9 +12,8 @@ use frame_support::{
 use frame_system::EnsureRoot;
 use orml_xcm_support::IsNativeConcrete;
 use pallet_xcm::XcmPassthrough;
-use polkadot_parachain::primitives::IsSystem;
-use polkadot_parachain::primitives::Sibling;
-use primitives::{currency::MultiCurrencyAdapter, tokens::*, Balance, ParaId};
+use polkadot_parachain::primitives::{IsSystem, Sibling};
+use primitives::{tokens::*, Balance, ParaId};
 use sp_core::H256;
 use sp_runtime::{
     generic,
@@ -33,14 +32,16 @@ pub use xcm_builder::{
     RelayChainAsNative, SiblingParachainAsNative, SiblingParachainConvertsVia,
     SignedAccountId32AsNative, SignedToAccountId32, SovereignSignedViaLocation, TakeWeightCredit,
 };
-use xcm_executor::traits::ConvertOrigin;
-use xcm_executor::{Config, XcmExecutor};
+use xcm_executor::{traits::ConvertOrigin, Config, XcmExecutor};
 use xcm_simulator::{decl_test_network, decl_test_parachain, decl_test_relay_chain};
 
 pub type AccountId = AccountId32;
 pub type CurrencyId = u32;
 pub use kusama_runtime;
-use pallet_traits::ump::{XcmCall, XcmWeightFeeMisc};
+use pallet_traits::{
+    ump::{XcmCall, XcmWeightFeeMisc},
+    xcm::MultiCurrencyAdapter,
+};
 
 pub struct RelayChainBlockNumberProvider<T>(sp_std::marker::PhantomData<T>);
 
@@ -348,9 +349,6 @@ impl SortedMembers<AccountId> for BobOrigin {
 }
 
 parameter_types! {
-    pub const CrowdloansPalletId: PalletId = PalletId(*b"crwloans");
-    pub const MinContribution: Balance = 0;
-    pub const MigrateKeysLimit: u32 = 10;
     pub SelfParaId: ParaId = para_a_id();
 }
