+++ primitives/Cargo.toml
@@ -18,9 +18,6 @@ sp-core                 = { git = 'https://github.com/paritytech/substrate.git',
 sp-io                   = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.17', default-features = false }
 sp-runtime              = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.17', default-features = false }
 sp-std                  = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.17', default-features = false }
-xcm                     = { git = 'https://github.com/paritytech/polkadot.git', branch = 'release-v0.9.17', default-features = false }
-xcm-builder             = { git = 'https://github.com/paritytech/polkadot.git', branch = 'release-v0.9.17', default-features = false }
-xcm-executor            = { git = 'https://github.com/paritytech/polkadot.git', branch = 'release-v0.9.17', default-features = false }
 
 [features]
 default = ['std']
@@ -35,9 +32,6 @@ std     = [
   'sp-consensus-aura/std',
   'frame-system/std',
   'frame-support/std',
-  'xcm-executor/std',
-  'xcm/std',
-  'xcm-builder/std',
   'scale-info/std',
   'cumulus-primitives-core/std',
 ]
